requirements.txt
django
djangorestframework
django-debug-toolbar
drf-nested-routers
drf-yasg
djoser
djangorestframework-simplejwt
pytest
pytest-django
